Moved old code to the new Linked Routine stack.
