+++ app/src/main/java/io/mazur/fit/adapter/RoutineAdapter.java
@@ -8,8 +8,11 @@ import android.widget.TextView;
 
 import butterknife.ButterKnife;
 import butterknife.InjectView;
+
 import io.mazur.fit.R;
 import io.mazur.fit.model.Routine;
+import io.mazur.fit.model.Exercise;
+import io.mazur.fit.model.LinkedRoutine;
 import io.mazur.fit.model.RoutineType;
 import io.mazur.fit.stream.RoutineStream;
 
@@ -35,7 +38,7 @@ public class RoutineAdapter extends RecyclerView.Adapter<RoutineAdapter.RoutineP
 
     @Override
     public void onBindViewHolder(RoutinePresenter holder, int position) {
-        holder.onBindView(mRoutine.getPartRoutines().get(position));
+        holder.onBindView(mRoutine.getLinkedRoutine().get(position));
     }
 
     @Override
@@ -44,12 +47,12 @@ public class RoutineAdapter extends RecyclerView.Adapter<RoutineAdapter.RoutineP
             return 0;
         }
 
-        return mRoutine.getSize();
+        return mRoutine.getLinkedRoutine().size();
     }
 
     @Override
     public int getItemViewType(int position) {
-        return mRoutine.getPartRoutines().get(position).getType().ordinal();
+        return mRoutine.getLinkedRoutine().get(position).getType().ordinal();
     }
 
     public void setRoutine(Routine routine) {
@@ -65,7 +68,7 @@ public class RoutineAdapter extends RecyclerView.Adapter<RoutineAdapter.RoutineP
             ButterKnife.inject(this, itemView);
         }
 
-        public abstract void onBindView(Routine.PartRoutine partRoutine);
+        public abstract void onBindView(LinkedRoutine linkedRoutine);
     }
 
     public class RoutineCategoryPresenter extends RoutinePresenter {
@@ -76,8 +79,8 @@ public class RoutineAdapter extends RecyclerView.Adapter<RoutineAdapter.RoutineP
         }
 
         @Override
-        public void onBindView(Routine.PartRoutine partRoutine) {
-            mRoutineCategoryTextView.setText(partRoutine.getTitle());
+        public void onBindView(LinkedRoutine linkedRoutine) {
+            mRoutineCategoryTextView.setText(linkedRoutine.getTitle());
         }
     }
 
@@ -89,8 +92,8 @@ public class RoutineAdapter extends RecyclerView.Adapter<RoutineAdapter.RoutineP
         }
 
         @Override
-        public void onBindView(Routine.PartRoutine partRoutine) {
-            mRoutineSectionTextView.setText(partRoutine.getTitle());
+        public void onBindView(LinkedRoutine linkedRoutine) {
+            mRoutineSectionTextView.setText(linkedRoutine.getTitle());
         }
     }
 
@@ -102,23 +105,23 @@ public class RoutineAdapter extends RecyclerView.Adapter<RoutineAdapter.RoutineP
         }
 
         @Override
-        public void onBindView(Routine.PartRoutine partRoutine) {
-            mRoutineExerciseTextView.setText(partRoutine.getTitle());
+        public void onBindView(LinkedRoutine linkedRoutine) {
+            mRoutineExerciseTextView.setText(linkedRoutine.getTitle());
 
-            itemView.setOnClickListener(new OnItemClickListener(partRoutine));
+            itemView.setOnClickListener(new OnItemClickListener(linkedRoutine));
         }
     }
 
     public class OnItemClickListener implements View.OnClickListener {
-        private Routine.PartRoutine mPartRoutine;
+        private LinkedRoutine mLinkedRoutine;
 
-        public OnItemClickListener(Routine.PartRoutine partRoutine) {
-            mPartRoutine = partRoutine;
+        public OnItemClickListener(LinkedRoutine linkedRoutine) {
+            mLinkedRoutine = linkedRoutine;
         }
 
         @Override
         public void onClick(View v) {
-            RoutineStream.getInstance().setExercise(mPartRoutine);
+            RoutineStream.getInstance().setExercise((Exercise) mLinkedRoutine);
         }
     }
 }
