+++ app/src/main/java/io/mazur/fit/model/LinkedRoutine.java
@@ -1,3 +1,6 @@
 package io.mazur.fit.model;
 
-public abstract  class LinkedRoutine {}
+public abstract class LinkedRoutine {
+    public abstract String getTitle();
+    public abstract RoutineType getType();
+}
