+++ app/src/main/java/io/mazur/fit/model/Section.java
@@ -20,6 +20,10 @@ public class Section extends LinkedRoutine {
         return mTitle;
     }
 
+    public RoutineType getType() {
+        return RoutineType.SECTION;
+    }
+
     public SectionMode getSectionMode() {
         return mSectionMode;
     }
