+++ app/src/main/java/io/mazur/fit/presenter/ListPresenter.java
@@ -2,7 +2,6 @@ package io.mazur.fit.presenter;
 
 import io.mazur.fit.adapter.RoutineAdapter;
 import io.mazur.fit.stream.RoutineStream;
-import io.mazur.fit.utils.Logger;
 import io.mazur.fit.view.ListView;
 
 public class ListPresenter {
@@ -21,10 +20,5 @@ public class ListPresenter {
 
         RoutineStream.getInstance().getRoutineObservable()
                 .subscribe(routine -> mRoutineAdapter.setRoutine(routine));
-
-        RoutineStream.getInstance().getExerciseChangedObservable().subscribe(exercise -> {
-            Logger.d("Scrolling to position: " + exercise.getSectionPosition());
-//            mListView.getRecyclerView().smoothScrollToPosition(exercise.getSectionPosition());
-        });
     }
 }
