+++ app/src/main/java/io/mazur/fit/stream/RoutineStream.java
@@ -9,7 +9,8 @@ import java.io.IOException;
 import io.mazur.fit.App;
 import io.mazur.fit.R;
 import io.mazur.fit.model.Routine;
-import io.mazur.fit.model.RoutineType;
+import io.mazur.fit.model.Exercise;
+import io.mazur.fit.model.JSONRoutine;
 import io.mazur.fit.utils.Logger;
 
 import rx.Observable;
@@ -21,50 +22,23 @@ public class RoutineStream {
     private static RoutineStream sInstance;
 
     private Routine mRoutine;
-    private Routine.PartRoutine mExercise;
+    private Exercise mExercise;
 
     private final PublishSubject<Routine> mRoutineSubject = PublishSubject.create();
-    private final PublishSubject<Routine.PartRoutine> mExerciseSubject = PublishSubject.create();
+    private final PublishSubject<Exercise> mExerciseSubject = PublishSubject.create();
 
     private RoutineStream() {
         try {
-            mRoutine = new Gson().fromJson(IOUtils.toString(App.getContext()
+            /**
+             * This should be read from Glacier if exists.
+             */
+            JSONRoutine jsonRoutine = new Gson().fromJson(IOUtils.toString(App.getContext()
                     .getResources()
                     .openRawResource(R.raw.beginner_routine)
-            ), Routine.class);
+            ), JSONRoutine.class);
 
-            /**
-             * TODO: Link Routine, I don't like how this is done at the moment.
-             * TODO: There is definitely a better way.
-             *
-             * Allows to link objects inside the Array List together so we can have previous
-             * and next buttons inside the application.
-             */
-            Routine.PartRoutine prev = null;
-            int index = 0, currentSectionPosition = 0;
-
-            for(Routine.PartRoutine partRoutine : mRoutine.getPartRoutines()) {
-                if(partRoutine.getType() == RoutineType.SECTION) {
-                    currentSectionPosition = index;
-                }
-
-                if(partRoutine.getType() == RoutineType.EXERCISE) {
-
-
-                    if(prev != null) {
-                        partRoutine.setSectionPosition(currentSectionPosition);
-                        partRoutine.setPrevious(prev);
-
-                        prev.setNext(partRoutine);
-                    }
-
-                    prev = partRoutine;
-                }
-
-                index++;
-            }
-
-            mExercise = mRoutine.getFirstExercise();
+            mRoutine = new Routine(jsonRoutine);
+            mExercise = mRoutine.getLinkedExercises().get(0);
 
             mRoutineSubject.onNext(mRoutine);
             mExerciseSubject.onNext(mExercise);
@@ -95,22 +69,22 @@ public class RoutineStream {
         return Observable.merge(mRoutineSubject, routineObservable);
     }
 
-    public void setExercise(Routine.PartRoutine exercise) {
+    public void setExercise(Exercise exercise) {
         mExercise = exercise;
         mExerciseSubject.onNext(exercise);
     }
 
-    public Observable<Routine.PartRoutine> getExerciseChangedObservable() {
+    public Observable<Exercise> getExerciseChangedObservable() {
         return mExerciseSubject;
     }
 
     /**
      * @return Observable that allows to subscribe when only exercise has changed.
      */
-    public Observable<Routine.PartRoutine> getExerciseObservable() {
-        Observable<Routine.PartRoutine> exerciseObservable = Observable.create(new Observable.OnSubscribe<Routine.PartRoutine>() {
+    public Observable<Exercise> getExerciseObservable() {
+        Observable<Exercise> exerciseObservable = Observable.create(new Observable.OnSubscribe<Exercise>() {
             @Override
-            public void call(Subscriber<? super Routine.PartRoutine> subscriber) {
+            public void call(Subscriber<? super Exercise> subscriber) {
                 subscriber.onNext(mExercise);
             }
         }).observeOn(AndroidSchedulers.mainThread()).publish().refCount();
